+++ internal/aws/applicationinsights/application_gen.go
@@ -64,6 +64,18 @@ func applicationResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 			//   "items": {
 			//     "additionalProperties": false,
 			//     "description": "The monitoring setting of the component.",
+			//     "oneOf": [
+			//       {
+			//         "required": [
+			//           "ComponentName"
+			//         ]
+			//       },
+			//       {
+			//         "required": [
+			//           "ComponentARN"
+			//         ]
+			//       }
+			//     ],
 			//     "properties": {
 			//       "ComponentARN": {
 			//         "description": "The ARN of the compnonent.",
@@ -1507,6 +1519,18 @@ func applicationResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 				},
 			),
 			Optional: true,
+			Validators: []tfsdk.AttributeValidator{
+				validate.RequiredAttributes(
+					validate.OneOfRequired(
+						validate.Required(
+							"component_name",
+						),
+						validate.Required(
+							"component_arn",
+						),
+					),
+				),
+			},
 		},
 		"custom_components": {
 			// Property: CustomComponents