+++ internal/aws/imagebuilder/image_gen.go
@@ -219,6 +219,18 @@ func imageResourceType(ctx context.Context) (tfsdk.ResourceType, error) {
 
 	opts = opts.WithUpdateTimeoutInMinutes(0)
 
+	opts = opts.WithRequiredAttributesValidators(validate.OneOfRequired(
+		validate.Required(
+			"container_recipe_arn",
+			"infrastructure_configuration_arn",
+		),
+		validate.Required(
+			"image_recipe_arn",
+			"infrastructure_configuration_arn",
+		),
+	),
+	)
+
 	resourceType, err := NewResourceType(ctx, opts...)
 
 	if err != nil {