+++ internal/aws/ssmcontacts/contact_channel_gen.go
@@ -149,6 +149,16 @@ func contactChannelResourceType(ctx context.Context) (tfsdk.ResourceType, error)
 
 	opts = opts.WithUpdateTimeoutInMinutes(0)
 
+	opts = opts.WithRequiredAttributesValidators(validate.OneOfRequired(
+		validate.Required(
+			"contact_id",
+			"channel_name",
+			"channel_type",
+			"channel_address",
+		),
+	),
+	)
+
 	resourceType, err := NewResourceType(ctx, opts...)
 
 	if err != nil {