+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## 0.1.0 (Unreleased)
 
+## [0.0.5](https://github.com/hashicorp/terraform-provider-awscc/releases/tag/v0.0.5) (August 31, 2021)
+
+Attribute validation support added.
+Additional AWS authnentication methods supported.
+
 ## [0.0.4](https://github.com/hashicorp/terraform-provider-awscc/releases/tag/v0.0.4) (August 19, 2021)
 
 Provider renamed to `terraform-provider-awscc`.